git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@973 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/subnetcalc.cc
@@ -140,11 +140,11 @@ int main(int argc, char** argv)
    }
    wildcard  = ~netmask;
 
-   printf("Network       = "); putAddress(network); printf(" / %u\n", prefix);
+   printf("Network       = "); putAddress(network); printf(" / %u\n", (unsigned long)prefix);
    printf("Netmask       = "); putAddress(netmask); puts("");
    printf("Broadcast     = "); putAddress(broadcast); puts("");
    printf("Wildcard Mask = "); putAddress(wildcard); puts("");
-   printf("Max. Hosts    = %u\n", hosts);
+   printf("Max. Hosts    = %u\n", (unsigned long)hosts);
    printf("Host Range    = { ");
    putAddress(host1); printf(" ... "); putAddress(host2);
    puts(" }");
